+++ src/main/java/com/sparta/temueats/cart/controller/CartController.java
@@ -6,6 +6,7 @@ import com.sparta.temueats.cart.dto.CartGetListResponseDto;
 import com.sparta.temueats.cart.service.CartService;
 import com.sparta.temueats.global.ResponseDto;
 import com.sparta.temueats.global.ex.CustomApiException;
+import com.sparta.temueats.security.UserDetailsImpl;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -20,22 +21,20 @@ import java.util.UUID;
 @RequestMapping("/api")
 public class CartController {
 
-    static final Long USER_ID = 1L;
-
     private final CartService cartService;
 
     // 장바구니 추가
     @PostMapping("/carts/{menuId}")
     public ResponseDto<?> cartAddResponseDto(@RequestBody @Valid CartUpdateRequestDto cartUpdateRequestDto, BindingResult bindingResult,
-                                             @PathVariable UUID menuId) {
-        CartUpdateResponseDto cartUpdateResponseDto = cartService.createCarts(cartUpdateRequestDto, USER_ID, menuId);
+                                             @PathVariable UUID menuId, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        CartUpdateResponseDto cartUpdateResponseDto = cartService.createCarts(cartUpdateRequestDto, userDetails.getUser(), menuId);
         return new ResponseDto<>(1, "장바구니 추가가 완료되었습니다.", cartUpdateResponseDto);
     }
 
     // 장바구니 전체 조회
     @GetMapping("/carts")
-    public ResponseDto<?> cartGetListResponseDto() {
-        List<CartGetListResponseDto> cartGetListResponseListDto = cartService.getCarts();
+    public ResponseDto<?> cartGetListResponseDto(@AuthenticationPrincipal UserDetailsImpl userDetails) {
+        List<CartGetListResponseDto> cartGetListResponseListDto = cartService.getCarts(userDetails.getUser());
         return new ResponseDto<>(1, "장바구니 전체 조회가 완료되었습니다.", cartGetListResponseListDto);
 
     }
@@ -44,21 +43,21 @@ public class CartController {
     @PutMapping("/carts/{cartId}/modify")
     public ResponseDto<?> cartUpdateResponseDto(@RequestBody @Valid CartUpdateRequestDto cartUpdateRequestDto, BindingResult bindingResult,
                                                 @PathVariable UUID cartId) {
-        CartUpdateResponseDto cartUpdateResponseDto = cartService.updateCarts(cartUpdateRequestDto, USER_ID, cartId);
+        CartUpdateResponseDto cartUpdateResponseDto = cartService.updateCarts(cartUpdateRequestDto, cartId);
         return new ResponseDto<>(1, "장바구니 메뉴 수량 변경이 완료되었습니다.", cartUpdateResponseDto);
     }
 
     // 장바구니 삭제
     @PutMapping("/carts/{cartId}/delete")
     public ResponseDto<?> cartDeleteResponse(@PathVariable UUID cartId) {
-        cartService.deleteCarts(USER_ID, cartId);
+        cartService.deleteCarts(cartId);
         return new ResponseDto<>(1, "장바구니 메뉴가 정상적으로 삭제되었습니다.", null);
     }
 
     // 장바구니 결제할 물품 선택
     @PutMapping("/carts/{cartId}/select")
     public ResponseDto<?> cartSelectResponse(@PathVariable UUID cartId) {
-        cartService.selectCarts(USER_ID, cartId);
+        cartService.selectCarts(cartId);
         return new ResponseDto<>(1, "장바구니 메뉴 선택이 수정되었습니다.", null);
     }
