+++ src/main/java/com/sparta/temueats/cart/service/CartService.java
@@ -6,11 +6,8 @@ import com.sparta.temueats.cart.dto.CartGetListResponseDto;
 import com.sparta.temueats.cart.entity.P_cart;
 import com.sparta.temueats.cart.repository.CartRepository;
 import com.sparta.temueats.global.ex.CustomApiException;
-import com.sparta.temueats.menu.entity.Category;
 import com.sparta.temueats.menu.entity.P_menu;
 import com.sparta.temueats.menu.repository.MenuRepository;
-import com.sparta.temueats.store.entity.P_store;
-import com.sparta.temueats.store.entity.SellState;
 import com.sparta.temueats.user.entity.P_user;
 import com.sparta.temueats.user.entity.UserRoleEnum;
 import lombok.RequiredArgsConstructor;
@@ -22,57 +19,25 @@ import org.springframework.transaction.annotation.Transactional;
 import java.sql.Date;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
 public class CartService {
-    // todo 더미 데이터 싹 수정
-
-    static GeometryFactory geometryFactory = new GeometryFactory();
-    static P_user staticCustomerUser = P_user.builder()
-            .id(4L)
-            .email("email@naver.com")
-            .password("임시1234")
-            .phone("010-1234-5678")
-            .nickname("네임")
-            .birth(Date.valueOf("2002-12-26"))
-            .use_yn(true)
-            .role(UserRoleEnum.OWNER)
-            .imageProfile("img_url")
-            .latLng(geometryFactory.createPoint(new Coordinate(123, 123)))
-            .address("123층 123호")
-            .build();
-
-    static P_user staticOwnerUser = P_user.builder()
-            .id(3L)
-            .email("email5@naver.com")
-            .password("임시12345")
-            .phone("010-5555-5555")
-            .nickname("네임5")
-            .birth(Date.valueOf("2002-12-26"))
-            .use_yn(true)
-            .role(UserRoleEnum.OWNER)
-            .imageProfile("img_url")
-            .latLng(geometryFactory.createPoint(new Coordinate(55, 55)))
-            .address("555층 555호")
-            .build();
-    static P_user usedUser = staticOwnerUser;
 
     private final CartRepository cartRepository;
     private final MenuRepository menuRepository;
 
-    public CartUpdateResponseDto createCarts(CartUpdateRequestDto cartUpdateRequestDto, Long user, UUID menuId) {
+    public CartUpdateResponseDto createCarts(CartUpdateRequestDto cartUpdateRequestDto, P_user user, UUID menuId) {
         // 1. 장바구니가 비어 있으면 바로 담기 & 바로 리턴
-        List<P_cart> allByUserId = cartRepository.findAllByUserId(usedUser.getId());
+        List<P_cart> allByUserId = cartRepository.findAllByUserId(user.getId());
         P_menu menu = menuRepository.findById(menuId).orElseThrow(() ->
                 new CustomApiException("해당 메뉴가 존재하지 않습니다."));
         if (allByUserId.isEmpty()) {
             P_cart cart3 = cartRepository.save(P_cart.builder()
                     .quantity(cartUpdateRequestDto.getQuantity())
                     .selectYn(false)
-                    .user(usedUser)
+                    .user(user)
                     .menu(menu)
                     .deletedYn(false)
                     .build());
@@ -80,7 +45,7 @@ public class CartService {
         }
 
         // 2-1. 장바구니에 같은 메뉴 id가 있는 경우
-        P_cart cart = cartRepository.findByMenuIdByUserId(menu.getMenuId(), usedUser.getId());
+        P_cart cart = cartRepository.findByMenuIdByUserId(menu.getMenuId(), user.getId());
         if (cart != null) {
             // 이미 같은 메뉴가 있다는 예외
             throw new CustomApiException("이미 해당 메뉴가 장바구니에 존재합니다.");
@@ -97,15 +62,15 @@ public class CartService {
         P_cart cart3 = cartRepository.save(P_cart.builder()
                         .quantity(cartUpdateRequestDto.getQuantity())
                         .selectYn(false)
-                        .user(usedUser)
+                        .user(user)
                         .menu(menu)
                         .deletedYn(false)
                 .build());
         return new CartUpdateResponseDto(cart3);
     }
 
-    public List<CartGetListResponseDto> getCarts() {
-        List<P_cart> cartList = cartRepository.findAllByUserId(usedUser.getId());
+    public List<CartGetListResponseDto> getCarts(P_user user) {
+        List<P_cart> cartList = cartRepository.findAllByUserId(user.getId());
         List<CartGetListResponseDto> responseDtoList = new ArrayList<>();
 
         for (P_cart cart : cartList) {
@@ -116,7 +81,7 @@ public class CartService {
     }
 
     @Transactional
-    public CartUpdateResponseDto updateCarts(CartUpdateRequestDto cartUpdateRequestDto, Long user, UUID cartId) {
+    public CartUpdateResponseDto updateCarts(CartUpdateRequestDto cartUpdateRequestDto, UUID cartId) {
         Long updatePrice = cartUpdateRequestDto.getQuantity();
 
         if (updatePrice <= 0 || updatePrice > 50) {
@@ -131,14 +96,14 @@ public class CartService {
     }
 
     @Transactional
-    public void deleteCarts(Long user, UUID cartId) {
+    public void deleteCarts(UUID cartId) {
         P_cart cart = cartRepository.findById(cartId).orElseThrow(() ->
                 new CustomApiException("해당 장바구니 품목을 찾을 수 없습니다."));
         cart.delete();
     }
 
     @Transactional
-    public void selectCarts(Long user, UUID cartId) {
+    public void selectCarts(UUID cartId) {
         P_cart cart = cartRepository.findById(cartId).orElseThrow(() ->
                 new CustomApiException("해당 장바구니 품목을 찾을 수 없습니다."));
         cart.changeSelect();
