+++ src/main/java/com/sparta/temueats/order/controller/OrderController.java
@@ -4,9 +4,14 @@ import com.sparta.temueats.global.ResponseDto;
 import com.sparta.temueats.order.dto.DeliveryOrderCreateRequestDto;
 import com.sparta.temueats.order.dto.OrderGetResponseDto;
 import com.sparta.temueats.order.dto.TakeOutOrderCreateRequestDto;
+import com.sparta.temueats.order.service.OrderCustomerService;
+import com.sparta.temueats.order.service.OrderOwnerService;
 import com.sparta.temueats.order.service.OrderService;
+import com.sparta.temueats.security.UserDetailsImpl;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
@@ -18,42 +23,49 @@ import java.util.UUID;
 @RequestMapping("/api")
 public class OrderController {
 
-    static final Long USER_ID = 1L;
-
     private final OrderService orderService;
+    private final OrderCustomerService orderCustomerService;
+    private final OrderOwnerService orderOwnerService;
 
     // todo customer 접근 가능 권한 설정
     // todo owner 접근 가능 권한 설정
 
     // 주문 생성 (Customer)
+    @PreAuthorize("hasAnyAuthority('CUSTOMER', 'MASTER')")
     @PostMapping("/orders/customer")
-    public ResponseDto<?> orderCreateResponseCustomerDto(@RequestBody @Valid DeliveryOrderCreateRequestDto deliveryOrderCreateRequestDto, BindingResult bindingResult) {
-        orderService.createDeliveryOrders(deliveryOrderCreateRequestDto, USER_ID);
+    public ResponseDto<?> orderCreateResponseCustomerDto(@RequestBody @Valid DeliveryOrderCreateRequestDto deliveryOrderCreateRequestDto, BindingResult bindingResult,
+                                                         @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        orderCustomerService.createDeliveryOrders(deliveryOrderCreateRequestDto, userDetails.getUser());
         return new ResponseDto<>(1, "주문이 정상적으로 생성되었습니다.", null);
     }
 
     // 주문 생성 (Owner)
+    @PreAuthorize("hasAnyAuthority('OWNER', 'MASTER')")
     @PostMapping("/orders/owner")
-    public ResponseDto<?> orderCreateResponseOwnerDto(@RequestBody @Valid TakeOutOrderCreateRequestDto takeOutOrderCreateRequestDto, BindingResult bindingResult) {
-        orderService.createTakeOutOrders(takeOutOrderCreateRequestDto, USER_ID);
+    public ResponseDto<?> orderCreateResponseOwnerDto(@RequestBody @Valid TakeOutOrderCreateRequestDto takeOutOrderCreateRequestDto, BindingResult bindingResult,
+                                                      @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        orderOwnerService.createTakeOutOrders(takeOutOrderCreateRequestDto, userDetails.getUser());
         return new ResponseDto<>(1, "주문이 정상적으로 생성되었습니다.", null);
     }
 
     // 주문 목록 조회 (Customer)
+    @PreAuthorize("hasAnyAuthority('CUSTOMER', 'MASTER')")
      @GetMapping("/orders/customer")
-    public ResponseDto<?> orderGetCustomerResponseListDto() {
-         List<OrderGetResponseDto> customerOrders = orderService.getCustomerOrders();
+    public ResponseDto<?> orderGetCustomerResponseListDto(@AuthenticationPrincipal UserDetailsImpl userDetails) {
+         List<OrderGetResponseDto> customerOrders = orderCustomerService.getCustomerOrders(userDetails.getUser());
          return new ResponseDto<>(1, "주문 목록 조회에 성공했습니다.", customerOrders);
      }
 
     // 주문 목록 조회 (Owner)
+    @PreAuthorize("hasAnyAuthority('OWNER', 'MASTER')")
      @GetMapping("/orders/owner")
-    public ResponseDto<?> orderGetOwnerResponseListDto() {
-        List<OrderGetResponseDto> ownerOrders = orderService.getOwnerOrders();
+    public ResponseDto<?> orderGetOwnerResponseListDto(@AuthenticationPrincipal UserDetailsImpl userDetails) {
+        List<OrderGetResponseDto> ownerOrders = orderOwnerService.getOwnerOrders(userDetails.getUser());
         return new ResponseDto<>(1, "주문 목록 조회에 성공했습니다.", ownerOrders);
      }
 
     // 주문 상세 조회 (Customer)
+    @PreAuthorize("hasAnyAuthority('CUSTOMER', 'MASTER')")
     @GetMapping("/orders/customer/{orderId}")
     public ResponseDto<?> orderGetCustomerResponseDto(@PathVariable UUID orderId) {
         OrderGetResponseDto customerOrder = orderService.getOrder(orderId);
@@ -61,6 +73,7 @@ public class OrderController {
     }
 
     // 주문 상세 조회 (Owner)
+    @PreAuthorize("hasAnyAuthority('OWNER', 'MASTER')")
     @GetMapping("/orders/owner/{orderId}")
     public ResponseDto<?> orderGetOwnerResponseDto(@PathVariable UUID orderId) {
         OrderGetResponseDto ownerOrder = orderService.getOrder(orderId);
@@ -69,16 +82,18 @@ public class OrderController {
 
 
     // 주문 취소 (Customer)
+    @PreAuthorize("hasAnyAuthority('CUSTOMER', 'MASTER')")
     @PutMapping("/orders/{orderId}/customer")
     public ResponseDto<?> orderCancelCustomerResponseDto(@PathVariable UUID orderId) {
-        orderService.cancelCustomerOrder(orderId);
+        orderCustomerService.cancelCustomerOrder(orderId);
         return new ResponseDto<>(1, "주문이 취소되었습니다.", null);
     }
 
     // 주문 취소 (Owner)
+    @PreAuthorize("hasAnyAuthority('OWNER', 'MASTER')")
     @PutMapping("/orders/{orderId}/owner")
     public ResponseDto<?> orderCancelOwnerResponseDto(@PathVariable UUID orderId) {
-        orderService.cancelOwnerOrder(orderId);
+        orderOwnerService.cancelOwnerOrder(orderId);
         return new ResponseDto<>(1, "주문이 취소되었습니다.", null);
     }
